# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FundCast API
  description: |
    ## AI-First Social Funding + Forecasting Platform for SaaS Founders
    
    FundCast is the premier platform connecting SaaS founders through prediction markets and social funding mechanisms. 
    Our enterprise-grade API provides comprehensive access to:
    
    - **🔮 Prediction Markets**: Binary, categorical, and scalar markets with dual-engine trading
    - **💰 Social Funding**: Reg CF and Rule 506(c) compliant investment workflows  
    - **🤖 AI Intelligence**: Semantic search, market analytics, and founder matching
    - **👥 Founder Network**: Elite community features with Purple Tier home screen featuring
    - **🛡️ Enterprise Security**: OWASP ASVS Level 2+ compliance with Red Team AI defense
    
    ### Authentication
    
    All API endpoints require authentication via JWT tokens:
    
    ```bash
    Authorization: Bearer <jwt_token>
    ```
    
    ### Rate Limiting
    
    - **Free Tier**: 60 requests/minute
    - **Oracle Tier**: 300 requests/minute  
    - **Whale Tier**: 1000 requests/minute
    - **Purple Tier**: 5000 requests/minute
    - **Kingmaker Tier**: Unlimited
    
    ### API Versioning
    
    We use URL path versioning with semantic versioning principles:
    - **Current**: `/api/v1/`
    - **Beta**: `/api/v2-beta/` (Preview features)
    
    ### Error Handling
    
    All errors follow RFC 7807 Problem Details format:
    
    ```json
    {
...
      "status": 422,
...
    }
    ```
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /markets:
    get:
      tags: [Markets]
      summary: List prediction markets
      description: Retrieve list of prediction markets with filtering and pagination
      operationId: listMarkets
      parameters:
        - name: category
          in: query
          description: Filter by market category
          schema:
            type: string
            enum: [technology, business, funding, founders, general]
        - name: status
          in: query
          description: Filter by market status
          schema:
            type: string
            enum: [active, resolved, cancelled]
        - name: limit
          in: query
          description: Number of markets to return (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of markets to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Markets retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketsListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
    
    post:
      tags: [Markets]
      summary: Create prediction market
      description: Create a new prediction market (Whale tier and above)
      operationId: createMarket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarketCreateRequest'
      responses:
        '201':
          description: Market created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Market'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /markets/{market_id}:
    get:
      tags: [Markets]
      summary: Get market details
      description: Retrieve detailed information about a specific market
      operationId: getMarketDetails
      parameters:
        - name: market_id
          in: path
          required: true
          description: Market UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Market details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketDetails'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # =============================================================================
  # Compliance
  # =============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    MarketsListResponse:
      type: object
      properties:
        markets:
          type: array
          items:
            $ref: '#/components/schemas/Market'
        total:
          type: integer
          description: Total number of markets matching filters
        limit:
          type: integer
        offset:
          type: integer
        has_next:
          type: boolean
    Market:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [technology, business, funding, founders, general]
        market_type:
          type: string
          enum: [binary, categorical, scalar]
        status:
          type: string
          enum: [active, paused, resolved, cancelled]
        created_by:
          type: string
          format: uuid
        resolution_date:
          type: string
          format: date-time
        current_probability:
          type: number
          format: float
          minimum: 0
          maximum: 1
        total_volume:
          type: integer
          description: Total trading volume in cents
        participant_count:
          type: integer
        created_at:
          type: string
          format: date-time
    MarketDetails:
      allOf:
        - $ref: '#/components/schemas/Market'
        - type: object
          properties:
            positions:
              type: array
              items:
                $ref: '#/components/schemas/MarketPosition'
            price_history:
              type: array
              items:
                $ref: '#/components/schemas/PricePoint'
            recent_trades:
              type: array
              items:
                $ref: '#/components/schemas/Trade'
    MarketCreateRequest:
      type: object
      required: [title, description, category, market_type, resolution_date]
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        category:
          type: string
          enum: [technology, business, funding, founders, general]
        market_type:
          type: string
          enum: [binary, categorical, scalar]
        resolution_date:
          type: string
          format: date-time
        resolution_source:
          type: string
          maxLength: 500
        initial_probability:
          type: number
          format: float
          minimum: 0.01
          maximum: 0.99
          default: 0.5
    MarketPosition:
      type: object
      properties:
        outcome:
          type: string
        shares:
          type: number
          format: float
        avg_price:
          type: number
          format: float
        current_value:
          type: integer
        unrealized_pnl:
          type: integer

    # =============================================================================
    # Compliance Schemas
    # =============================================================================
...
    PricePoint:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        price:
          type: number
          format: float
        volume:
          type: integer
    Trade:
      type: object
      properties:
        id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        outcome:
          type: string
        shares:
          type: number
          format: float
        price:
          type: number
          format: float
        trader_id:
          type: string
          format: uuid

    # =============================================================================
    # Error Schemas
    # =============================================================================
    Error:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type
        title:
          type: string
          description: Short, human-readable summary of the problem
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation of the problem
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string
  responses:
...
    UnauthorizedError:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://fundcast.ai/errors/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Valid authentication credentials are required"
            instance: "/api/v1/users/profile"
    ForbiddenError:
      description: Access denied - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://fundcast.ai/errors/forbidden"
            title: "Forbidden" 
            status: 403
            detail: "This feature requires Purple tier subscription or higher"
            instance: "/api/v1/subscriptions/purple-featuring/queue"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://fundcast.ai/errors/not-found"
            title: "Not Found"
            status: 404
            detail: "The requested resource could not be found"
            instance: "/api/v1/markets/123e4567-e89b-12d3-a456-426614174000"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            type: "https://fundcast.ai/errors/validation-failed"
            title: "Validation Failed"
            status: 422
            detail: "The request contains invalid parameters"
            instance: "/api/v1/auth/register"
            errors:
              - field: "password"
                message: "Password must contain at least one uppercase letter"
                code: "password_requirements"
...